# api/swagger.yaml
openapi: 3.0.3
info:
  title: DevNest Portal API (OpenAPI 3.0)
  description: The DevNest Portal API endpoints definitions based on the OpenAPI 3.0
    specification.
  termsOfService: https://devnest.ro/terms-and-conditions
  contact:
    email: contact@devnest.ro
  version: 1.0.0
...
servers:
- url: https://me.devnest.ro/api/v1
...
paths:
  /timesheets:
    get:
      tags:
      - Timesheets
      summary: Returns all timesheets
      description: Returns all timesheets
      operationId: getAllTimesheets
      parameters:
      - name: projectName
        in: query
        description: return all timesheets by projectName
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: limit
        in: query
        description: number of timesheets to return
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: int
          default: 5
      - name: date
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
            format: date
      responses:
        "200":
          description: Returns an array of timesheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreatedTimesheet'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /timesheets/{timesheetId}:
    get:
      tags:
      - Timesheets
      summary: Returns timesheet by timesheetId
      description: Return timesheet by timesheetId
      operationId: getTimesheetById
      parameters:
      - name: timesheetId
        in: path
        description: return timesheet by timesheetId
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      responses:
        "200":
          description: returns timesheet by timesheetId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedTimesheet'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
      - Timesheets
      summary: Update an timesheet
      description: Update an timesheet
      parameters:
      - name: timesheetId
        in: path
        description: ID of the timesheet
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      requestBody:
        description: Timesheet data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Timesheet'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timesheet'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /timesheets/project-status/{projectId}:
    get:
      tags:
      - Timesheets
      summary: Returns project-status from timesheet by projectId
      description: Returns project-status from timesheet by projectId
      operationId: getProjectStatusByProjectId
      parameters:
      - name: projectId
        in: path
        description: Id of the project
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      responses:
        "200":
          description: returns projectStatus from timesheet by projectId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedProjectStatus'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Timesheet:
      type: object
      properties:
        id:
          type: number
          format: int
        day1:
          type: number
          format: int
        day2:
          type: number
          format: int
        day3:
          type: number
          format: int
        day4:
          type: number
          format: int
        day5:
          type: number
          format: int
        day6:
          type: number
          format: int
        day7:
          type: number
          format: int
        day8:
          type: number
          format: int
        day9:
          type: number
          format: int
        day10:
          type: number
          format: int
        day11:
          type: number
          format: int
        day12:
          type: number
          format: int
        day13:
          type: number
          format: int
        day14:
          type: number
          format: int
        day15:
          type: number
          format: int
        day16:
          type: number
          format: int
        day17:
          type: number
          format: int
        day18:
          type: number
          format: int
        day19:
          type: number
          format: int
        day20:
          type: number
          format: int
        day21:
          type: number
          format: int
        day22:
          type: number
          format: int
        day23:
          type: number
          format: int
        day24:
          type: number
          format: int
        day25:
          type: number
          format: int
        day26:
          type: number
          format: int
        day27:
          type: number
          format: int
        day28:
          type: number
          format: int
        day29:
          type: number
          format: int
        day30:
          type: number
          format: int
        day31:
          type: number
          format: int
      example:
        day29: 2.884162126668780246063761296682059764862060546875
        day28: 1.2846590061165319429647979632136411964893341064453125
        day27: 6.96511769763884558415156789124011993408203125
        day26: 3.557195227068097320710649000830017030239105224609375
        day14: 1.489415909854170383397331534069962799549102783203125
        day13: 1.024645700144157789424070870154537260532379150390625
        day12: 1.231513536777255612975068288506008684635162353515625
        day11: 7.3862819483858839220147274318151175975799560546875
        day10: 4.1456080298839363962315474054776132106781005859375
        day31: 6.87805222012787620400331434211693704128265380859375
        day30: 6.77832496304801335185175048536621034145355224609375
        id: 0.80082819046101150206595775671303272247314453125
        day19: 5.02500479152029466689555192715488374233245849609375
        day18: 4.9652184929849543237878606305457651615142822265625
        day17: 1.173074250955943309548956676735542714595794677734375
        day16: 7.4577447736837658709418974467553198337554931640625
        day15: 6.8468526983526398765889098285697400569915771484375
        day25: 6.438423552598546706349225132726132869720458984375
        day24: 9.0183481860707832566959041287191212177276611328125
        day8: 3.61607674925191080461672754609026014804840087890625
        day23: 8.7620420127490010742121739895083010196685791015625
        day9: 2.027123023002321833274663731572218239307403564453125
        day22: 6.683562403749608193948006373830139636993408203125
        day6: 7.061401241503109105224211816675961017608642578125
        day21: 9.3693102714106686335071572102606296539306640625
        day7: 9.301444243932575517419536481611430644989013671875
        day20: 9.965781217890562260208753286860883235931396484375
        day4: 5.63737665663332876420099637471139430999755859375
        day5: 2.3021358869347654518833223846741020679473876953125
        day2: 1.46581298050294517310021547018550336360931396484375
        day3: 5.962133916683182377482808078639209270477294921875
        day1: 6.02745618307040320615897144307382404804229736328125
...
    CreatedTimesheet:
      required:
      - date
      - id
      - projectId
      - status
      type: object
      properties:
        id:
          type: number
          format: int
        status:
          type: string
          default: unsubmitted
          enum:
          - submitted
          - unsubmitted
        date:
          type: string
          format: date
        projectId:
          type: number
          format: int
        file:
          type: string
          format: binary
      example:
        date: 2000-01-23
        file: ""
        id: 0.80082819046101150206595775671303272247314453125
        projectId: 6.02745618307040320615897144307382404804229736328125
        status: unsubmitted
    CreatedProjectStatus:
      required:
      - employeeId
      - extraHours
      - id
      - totalHours
      type: object
      properties:
        id:
          type: number
          format: int
        projectId:
          type: number
          format: int
        employeeId:
          type: number
          format: int
        totalHours:
          type: number
          format: int
        extraHours:
          type: number
          format: int
        rateType:
          type: string
          enum:
          - HOUR
          - MONTH
        rateValue:
          type: number
          format: int
      example:
        rateValue: 2.3021358869347654518833223846741020679473876953125
        rateType: HOUR
        extraHours: 5.63737665663332876420099637471139430999755859375
        totalHours: 5.962133916683182377482808078639209270477294921875
        employeeId: 1.46581298050294517310021547018550336360931396484375
        id: 0.80082819046101150206595775671303272247314453125
        projectId: 6.02745618307040320615897144307382404804229736328125
...
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      example:
        code: code
        message: message
...
  securitySchemes:
    devnest_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://me.devnest.ro/oauth/auth
          scopes:
            read:timesheets: Read timesheets
            read:projects: Read projects
            write:projects: Update projects
            read:employees: Read employees
            write:employees: Update employees
            read:allocations: Read allocations
            write:allocations: Update allocations
            read:history: Read history
            write:history: Update history
            read:feedback: Read feedback
            write:feedback: Update feedback
            read:settings: Read settings
            write:settings: Write settings