# go/model_created_project_status.go
/*
 * DevNest Portal API (OpenAPI 3.0)
 *
 * The DevNest Portal API endpoints definitions based on the OpenAPI 3.0 specification.
 *
 * API version: 1.0.0
 * Contact: contact@devnest.ro
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type CreatedProjectStatus struct {

	Id float64 `json:"id"`

	ProjectId float64 `json:"projectId,omitempty"`

	EmployeeId float64 `json:"employeeId"`

	TotalHours float64 `json:"totalHours"`

	ExtraHours float64 `json:"extraHours"`

	RateType string `json:"rateType,omitempty"`

	RateValue float64 `json:"rateValue,omitempty"`
}