# go/model_timesheet.go
/*
 * DevNest Portal API (OpenAPI 3.0)
 *
 * The DevNest Portal API endpoints definitions based on the OpenAPI 3.0 specification.
 *
 * API version: 1.0.0
 * Contact: contact@devnest.ro
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Timesheet struct {

	Id float64 `json:"id,omitempty"`

	Day1 float64 `json:"day1,omitempty"`

	Day2 float64 `json:"day2,omitempty"`

	Day3 float64 `json:"day3,omitempty"`

	Day4 float64 `json:"day4,omitempty"`

	Day5 float64 `json:"day5,omitempty"`

	Day6 float64 `json:"day6,omitempty"`

	Day7 float64 `json:"day7,omitempty"`

	Day8 float64 `json:"day8,omitempty"`

	Day9 float64 `json:"day9,omitempty"`

	Day10 float64 `json:"day10,omitempty"`

	Day11 float64 `json:"day11,omitempty"`

	Day12 float64 `json:"day12,omitempty"`

	Day13 float64 `json:"day13,omitempty"`

	Day14 float64 `json:"day14,omitempty"`

	Day15 float64 `json:"day15,omitempty"`

	Day16 float64 `json:"day16,omitempty"`

	Day17 float64 `json:"day17,omitempty"`

	Day18 float64 `json:"day18,omitempty"`

	Day19 float64 `json:"day19,omitempty"`

	Day20 float64 `json:"day20,omitempty"`

	Day21 float64 `json:"day21,omitempty"`

	Day22 float64 `json:"day22,omitempty"`

	Day23 float64 `json:"day23,omitempty"`

	Day24 float64 `json:"day24,omitempty"`

	Day25 float64 `json:"day25,omitempty"`

	Day26 float64 `json:"day26,omitempty"`

	Day27 float64 `json:"day27,omitempty"`

	Day28 float64 `json:"day28,omitempty"`

	Day29 float64 `json:"day29,omitempty"`

	Day30 float64 `json:"day30,omitempty"`

	Day31 float64 `json:"day31,omitempty"`
}